refactor: move suiClient initialization to client-side lib
